Make `image` dependency optional
Introduce an X11Image type, which allows querying pixel values.
Screen::capture() now returns an X11Image instead of an RgbImage.

Add an optional feature called image-interop, which enables converting
an X11Image into an RgbImage. This feature is enabled by default, but
can be disabled if the user does not care about the functionality
provided by the image crate.

Add a new example called get-pixel, which demonstrates the bare usage of
X11Image.

Rename the old example to save-png, which demonstrates converting the
X11Image into an RgbImage, and saving it to a PNG file.